used the built in sqrt function and passed all tests

+++ darts/darts.c
@@ -2,8 +2,8 @@
 #include <stdio.h>
 
 u_int8_t score(coordinate_t dart) {
-  int radius = (int)sqrt((dart.x * dart.x) + (dart.y * dart.y));
-  printf("Radius is: %d, for dart x: %f, y: %f\n", radius, dart.x, dart.y);
+  float radius = sqrt((dart.x * dart.x) + (dart.y * dart.y));
+  printf("Radius is: %f, for dart x: %f, y: %f\n", radius, dart.x, dart.y);
   if (radius <= 1) {
     return 10;
   } else if (radius <= 5) {
